requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
python-dotenv>=1.0.0
bedrock-agentcore>=1.1.0
tavily-python>=0.5.0
